Validate snapshots

+++ src/main/groovy/com/github/franciscozuccala/library/tasks/UploadAarTask.groovy
@@ -27,8 +27,6 @@ class UploadAarTask extends AbstractGithubTask {
         def aarsFolder = new File(gitFolder, "$groupId/$artifactId/$version")
         if (!aarsFolder.exists()) {
             aarsFolder.mkdirs()
-        }else if (!version.contains("SNAPSHOT")){
-            throw new Exception("Aar already exists in the repository")
         }
 
         println("Created folder ${aarsFolder.name} for aars")
@@ -66,6 +64,10 @@ class UploadAarTask extends AbstractGithubTask {
             File newAar = new File(folder, "$aar.name")
             println("Adding file named: $aar.name to $folder.name")
             if (newAar.exists()) {
+                if (newAar.name.contains("SNAPSHOT")){
+                    println("Aar called: $newAar.name already exists in the repository")
+                    return
+                }
                 newAar.delete()
             }
